Fix SearchFooter.test
